Added a linux window using X11 and XCB, CLion is still working on linux so I used it

+++ IzumaEngine/src/Platform/Platform.h
@@ -24,7 +24,7 @@ namespace Izuma
 
         virtual double GetTime() = 0;
 
-        virtual void Sleep(double milliseconds) = 0;
+        virtual void Sleep(uint32_t milliseconds) = 0;
 
         static Platform* GetPlatform();
